requirements.txt
opencv-python
mediapipe==0.10.14
keras
keras-tuner
scikit-learn
numpy
pandas
matplotlib
seaborn
yellowbrick
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-extensions==3.2.1
protobuf
channels
daphne
cryptography==41.0.7
